[#21] tests and cleanup

+++ backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/integrationtest/EventEndpointTest.java
@@ -14,6 +14,7 @@ import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.EventDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.FileDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.NewsDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.SmallLocationDto;
+import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.TopTenEventDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.mapper.ArtistMapper;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.mapper.EventMapper;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Artist;
@@ -26,6 +27,7 @@ import at.ac.tuwien.sepm.groupphase.backend.entity.SeatSector;
 import at.ac.tuwien.sepm.groupphase.backend.entity.SeatTicket;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Sector;
 import at.ac.tuwien.sepm.groupphase.backend.entity.StandingSector;
+import at.ac.tuwien.sepm.groupphase.backend.entity.StandingTicket;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Ticket;
 import at.ac.tuwien.sepm.groupphase.backend.entity.TicketOrder;
 import at.ac.tuwien.sepm.groupphase.backend.entity.User;
@@ -40,6 +42,8 @@ import at.ac.tuwien.sepm.groupphase.backend.repository.ArtistRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.EventRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.LocationRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.OrderRepository;
+import at.ac.tuwien.sepm.groupphase.backend.repository.PerformanceRepository;
+import at.ac.tuwien.sepm.groupphase.backend.repository.TicketRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.UserRepository;
 import at.ac.tuwien.sepm.groupphase.backend.security.JwtTokenizer;
 import at.ac.tuwien.sepm.groupphase.backend.util.FileDtoDeserializer;
@@ -104,7 +108,6 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
     @Autowired
     private EventRepository eventRepository;
 
-    @Autowired
     private static ObjectMapper MAPPER;
 
     @Autowired
@@ -115,6 +118,12 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
 
     @Autowired
     private ArtistRepository artistRepository;
+
+    @Autowired
+    private TicketRepository ticketRepository;
+
+    @Autowired
+    private PerformanceRepository performanceRepository;
 
     @Autowired
     private LocationRepository locationRepository;
@@ -150,7 +159,9 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
 
     @BeforeEach
     public void setCreateEventDto() throws FileNotFoundException {
-        eventRepository.deleteAll();
+        if(eventRepository != null) {
+            eventRepository.deleteAll();
+        }
         createEventDto = new CreateEventDto();
 
         //generate location for event
@@ -237,8 +248,17 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
         createEventDto.setThumbnail(fileDto);
     }
 
-    @BeforeEach
     public void createEventAndPurchaseTicketOrder() {
+        if(eventRepository != null) {
+            eventRepository.deleteAll();
+        }
+        orderRepository.deleteAll();
+        locationRepository.deleteAll();
+        artistRepository.deleteAll();
+        userRepository.deleteAll();
+        ticketRepository.deleteAll();
+        performanceRepository.deleteAll();
+
         //generate location for event
         Location location = new Location();
         location.setName(LOCATION_NAME);
@@ -251,6 +271,7 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
         location.setAddress(address);
 
         Set<Sector> sectors = new HashSet<>();
+        List<StandingSector> sectorList = new ArrayList<>();
         for (int j = 0; j < STANDING_SEC_ROWS; j++) {
             StandingSector sector = new StandingSector();
             sector.setName(STANDING_SEC_NAME + j);
@@ -266,29 +287,17 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
             sector.setPoint2(point);
             sector.setLocation(location);
             sectors.add(sector);
-        }
-        for (int j = 0; j < SEAT_SEC_ROWS; j++) {
-            SeatSector sector = new SeatSector();
-            sector.setPrice(SEAT_SEC_PRICE);
-            sector.setSeatType(SeatType.values()[j]);
-            List<Seat> seats = new ArrayList<>();
-
-            for (int k = 0; k < SEAT_SEC_ROWS * SEA_SEC_COLS; k++) {
-                Point point = new Point();
-                point.setX(k);
-                point.setY(5 + j);
-                Seat seat = new Seat();
-                seat.setRow(j + 1);
-                seat.setColumn(k + 1);
-                seat.setPoint(point);
-                seats.add(seat);
-            }
-            sector.setSeats(seats);
-            sector.setLocation(location);
-            sectors.add(sector);
+            sectorList.add(sector);
         }
         location.setSectors(sectors);
         location = locationRepository.save(location);
+
+        //Generate Artists
+        Set<Artist> artists = new HashSet<>();
+        for (int i = 0; i < NUMBER_OF_ARTISTS; i++) {
+            Artist a = artistRepository.save(new Artist(ARTIST_NAME, ARTIST_DESCRIPTION));
+            artists.add(a);
+        }
 
         //Generate event
         Event event = new Event();
@@ -296,29 +305,29 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
         event.setDescription(EVENT_TEST_DESCRIPTION);
         event.setDuration(EVENT_TEST_DURATION);
         event.setCategory(EVENT_CATEGORY);
+        event.setThumbnail(null);
+        event.setArtists(artists);
+        event = eventRepository.save(event);
+
+
+        //Generate Performances
         List<Performance> performances = new ArrayList<>();
         for (int i = 0; i < NUMBER_OF_PERFORMANCES; i++) {
-            Performance perf = new Performance(LocalDateTime.now().plusDays(i), location, event);
+            Performance perf = new Performance(LocalDateTime.now().plusDays((i + 1) * 2), location, event);
+            perf = performanceRepository.save(perf);
             performances.add(perf);
         }
         event.setPerformances(performances);
 
-        Set<Artist> artists = new HashSet<>();
-        for (int i = 0; i < NUMBER_OF_ARTISTS; i++) {
-            Artist a = artistRepository.save(new Artist(ARTIST_NAME, ARTIST_DESCRIPTION));
-            artists.add(a);
-        }
-        event.setArtists(artists);
-        event.setThumbnail(null);
-
-        eventRepository.save(event);
-
- /*       List<Ticket> tickets = new ArrayList<>();
-        SeatTicket ticket = new SeatTicket();
+        //Generate Tickets
+        List<Ticket> tickets = new ArrayList<>();
+        StandingTicket ticket = new StandingTicket();
+        ticket.setPerformance(performances.get(1));
         ticket.setPerformanceId(performances.get(1).getId());
-        ticket.setSectorId(sectors.stream().iterator().next().getId());
-        ticket.setSeatId(sectors.stream().iterator().next().getId());
-        tickets.add(ticket);*/
+        ticket.setSector(sectorList.get(1));
+        ticket.setSectorId(sectorList.get(1).getId());
+        ticket = ticketRepository.save(ticket);
+        tickets.add(ticket);
 
         Faker faker = new Faker();
 
@@ -326,7 +335,8 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
             .address(address).password("ABC").email(faker.name().firstName() + "." + faker.name().lastName() + '@')
             .role(UserRole.CUSTOMER).status(UserStatus.OK).readNews(new HashSet<>()).failedLoginAttempts(0).build();
 
-        userRepository.save(user);
+        user = userRepository.save(user);
+
 
         TicketOrder order = new TicketOrder();
         order.setType(OrderType.PURCHASE);
@@ -334,7 +344,10 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
         order.setValidUntil(order.getDateTime().plusHours(1)); // VALID FOR 1 HOUR
         order.setUser(user);
         order.setUserId(user.getId());
-        orderRepository.save(order);
+        order.setTickets(tickets);
+        order = orderRepository.save(order);
+        ticket.setOrder(order);
+        ticket.setOrderId(order.getId());
     }
 
 
@@ -413,9 +426,11 @@ public class EventEndpointTest implements EventTestData, ArtistTestData, Locatio
             .andReturn();
 
         MockHttpServletResponse response = mvcResult.getResponse();
-        List<EventDto> eventDtos = Arrays.asList(MAPPER.readValue(response.getContentAsString(),EventDto[].class));
-        assertEquals(1,eventDtos.size());
-
+        List<TopTenEventDto> eventDtos = Arrays.asList(MAPPER.readValue(response.getContentAsString(),TopTenEventDto[].class));
+        assertAll(
+            () -> assertEquals(1,eventDtos.size()),
+            () -> assertEquals(1, eventDtos.get(0).getTicketCount())
+        );
     }
 
 }
